requirements.txt
google-auth
google-api-python-client
google-cloud-aiplatform
Flask
heic2png